fix an incorrect realloc usage
realloc reallocates the given area of memory (implicitly creating
objects in the destination area). It must be previously allocated
by std::malloc, std::calloc or std::realloc and not yet freed
with std::free, otherwise, the results are undefined.
https://en.cppreference.com/w/cpp/memory/c/realloc.html

If the allocated type is a non-array type, the allocation
function's name is operator new and the deallocation function's
name is operator delete.
https://eel.is/c++draft/expr.new#12.sentence-3
https://eel.is/c++draft/basic.stc.dynamic.deallocation